ai_crew/requirements.txt
crewai>=0.28.0
langchain-community>=0.0.10
python-dotenv>=1.0.0
requests>=2.31.0
PyGithub>=2.0.0
slack-sdk>=3.21.3
ollama>=0.1.5
langchain-core>=0.1.0
gitpython>=3.1.0
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
